import arnold-kevin-2004-01-11.tar.bz2
from https://sourceforge.net/projects/arnold.berlios/

$ sha1sum arnold-kevin-2004-01-11.tar.bz2
76d5dca4cbfd5fe2883fa9344bd5d26d8b86a78f  arnold-kevin-2004-01-11.tar.bz2

the contents of the archive are identical to arnsrc11012004.zip,
which ships only the files modified since the arnsrc04012004 release.
the only difference is an added file README.FIRST with the following
content, which i didn't check in.

"""
2004-01-11 Andreas Micklei

This is the current patched version of Arnold as released by Kevin
Thacker on 2004-01-11. It is repackaged for easy install on Linux but
not carefully tested. Expect an enhanced Linux release in the near
future.
"""