snapv3: initialize pointer
warning from gcc:
cpc/snapv3.c:244:7: warning: 'pOutputData' may be used uninitialized in this function [-Wmaybe-uninitialized]
       ++pOutputData;

this is due to the default label in the switch statement.
maybe it's unreachable, but if it is it's better we crash with a
NULL-deref than using overwriting/reading random memory which could
lead to hard-to-find bugs.