change method to acquire rom size
the technique used in Makefile to embed the roms into the binary is
almost undocumented.

we do:
$(CC) -Wl,-b,binary,$(ROMS_BIN),-b,$(ELF_FMT)

were ROMS_BIN is a comma-separated list of filenames in the roms directory
and ELF_FMT is a string like elf64-x86-64, extracted from a call to
gcc -Wl,--print-output-format.

the only documentation about the -b,binary linker flag is actually in objcopy
documentation, which says:

> -B bfdarch --binary-architecture=bfdarch

> Useful when transforming a raw binary input file into an object file.
> In this case the output architecture can be set to bfdarch.
> This option will be ignored if the input file has a known bfdarch.
> You can access this binary data inside a program by referencing the
> special symbols that are created by the conversion process.
> These symbols are called _binary_objfile_start,
> _binary_objfile_end and _binary_objfile_size.
> e.g. you can transform a picture file into an object file and then access it
> in your code using these symbols.

however the documentation is misleading, as the actual syntax to achieve this
is:

    $ objcopy -I binary -O elf32-little binblob binblob.o
    $ file binblob.o
    binblob.o: ELF 32-bit LSB relocatable, no machine, version 1 (SYSV), not stripped
    $ nm  -S -t d binblob.o
    0000000001048576 D _binary_binblob_end
    0000000001048576 A _binary_binblob_size
    0000000000000000 D _binary_binblob_start

it's not even clear what the size of the _binary_foo_size member is (4 or 8?)
though it would be logical it being the native wordsize of the chosen elf
target.

when defining the size member as long, accessing the symbol crashes,
because it seems to contain the raw value rather than a relocation to the value.
therefore the confusing "(int) & size" code (while size being declared as int*)
was used, which threw pointer truncation warnings and caused crashes when the
binary was compiled as pie.

rather than dealing with this oddity, we now simply subtract start from end to
aquire the size of each binary blobb.

closes #3.